+++ src/ui/menu/WarpLoadPopup.hpp
@@ -5,7 +5,7 @@
 
 namespace globed {
 
-class WarpLoadPopup : public BasePopup<WarpLoadPopup, int, bool, bool>, public LevelDownloadDelegate, public LevelManagerDelegate {
+class WarpLoadPopup : public BasePopup<WarpLoadPopup, int, bool, bool>, public LevelDownloadDelegate {
 public:
     static const cocos2d::CCSize POPUP_SIZE;
     ~WarpLoadPopup();
@@ -14,14 +14,13 @@ private:
     cocos2d::CCLabelBMFont* m_statusLabel;
     int m_levelId;
     bool m_openLevel, m_replaceScene;
+    bool m_finished = false;
 
     bool setup(int levelId, bool openLevel, bool replaceScene) override;
     void onClose(cocos2d::CCObject*) override;
 
-    void loadLevelsFinished(cocos2d::CCArray* p0, char const* p1, int p2) override;
-    void loadLevelsFinished(cocos2d::CCArray* p0, char const* p1) override;
-    void loadLevelsFailed(char const* p0, int p1) override;
-    void loadLevelsFailed(char const* p0) override;
+    void onLoadedMeta(GJGameLevel* level);
+    void onLoadedData(GJGameLevel* level);
 
     void levelDownloadFinished(GJGameLevel* p0) override;
     void levelDownloadFailed(int p0) override;